requirements.txt
fastapi 
uvicorn
httpx
python-dotenv
pillow
aiosmtplib
openai
